nightswatcher: update and prune (#35)
Ubuntu 18.04, Python 3, and updated requirements.

Also use `openjdk-8-jre-headless` instead of `openjdk-8-jdk` which I'd stupidly used because I would on desktop.
The problem with `openjdk-8-jdk` is less that it *supports* audio and video, which would waste only some space, but more that it pulls in the likes of pulse, mesa and all their dependencies. The resulting image is some 200 MB smaller.